Enum collenchyma_nn::ConvBackwardDataAlgo

pub enum ConvBackwardDataAlgo {
    Auto,
    ImplicitGEMM,
    ImplicitGEMMSum,
    FFT,
    FFTTiling,
}

Different algorithms to compute the gradient with respect to the filter.

Variants

Auto

Attempt to automatically find the best algorithm of all the other available ones.

ImplicitGEMM

Compute the convolution as matrix product without forming the matrix that holds the input data.

Does not need any memory workspace.

The results are deterministic.

ImplicitGEMMSum

Compute the convolution as sum of matrix product without forming the matrix that holds the input data.

Does not need any memory workspace.

The results are non-deterministic.

FFT

Compute the convolution as Fast-Fourier Transform.

Needs a significant memory workspace.

The results are deterministic.

FFTTiling

Compute the convolution as Fast-Fourier Transform with 32x32 tiles.

Needs a significant memory workspace.

The results are deterministic.

Methods

impl ConvBackwardDataAlgo

fn is_auto(&self) -> bool

Check if algorithim should be chosen automatically.